[Refact] 쿼리문 분리
쿼리문이 분리되지않은 컴포넌트와 컨테이너의 쿼리문 분리.
