+++ web/src/queries/isLoggedIn.js
@@ -1,0 +1,9 @@
+import { gql } from 'apollo-boost';
+
+const IS_LOGGED_IN = gql`
+  query IsUserLoggedIn {
+    isLoggedIn @client
+  }
+`;
+
+export default IS_LOGGED_IN;
